backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
gitpython
tiktoken
chromadb
langchain
motor